build: コンパイル成功 (0 error, 41 files, 148KB DLL)
修正内容:
- HasTarget() → GetTarget().Index != 0
- IsEvacuating(Vector3) / EvacuateAll(bool) 正しいシグネチャ
- FindDisasterInfo<T> を PrefabCollection 検索に置換
- Notification.Problem → ProblemStruct.IsNone + ToString()
- PackageManager.GetAssetsOfType → 削除（未対応API）
- UpdateMode.LoadSimulation → NewGameFromMap/LoadGame
- MilestonesWrapper.instance → UnlockManager.instance.m_MilestonesWrapper
- JsonUtil.Number double→float キャスト
- uint→int 明示キャスト
- TransferCommands 未使用変数削除

+++ src/LoanCommands.cs
@@ -80,7 +80,7 @@ namespace SkylinesAgentBridge
                             json.Append("\"id\":" + i);
                             json.Append(",\"amount\":" + amount);
                             if (fInterest != null)
-                                json.Append(",\"interestRate\":" + JsonUtil.Number((int)fInterest.GetValue(loan) / 100.0));
+                                json.Append(",\"interestRate\":" + JsonUtil.Number((float)((int)fInterest.GetValue(loan) / 100.0)));
                             if (fWeekLength != null)
                                 json.Append(",\"weeksLeft\":" + (int)fWeekLength.GetValue(loan));
                             json.Append("}");
